scripts/requirements-migrate.txt
supabase>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
